fix big ID bug, process operation finish with partial test

+++ webStart/BackContent/src/main/java/backend/repository/IdManagerRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 /**
  * Created by lenovo on 2017/6/8.
  */
-public interface IdManagerRepository extends MongoRepository<IdManager, Integer> {
+public interface IdManagerRepository extends MongoRepository<IdManager, Long> {
 }
